function/requirements.txt
azure-functions==1.24.0
requests>=2.31.0
azure-storage-blob>=12.20.0

